requirements.txt
# ==============================
# Core Dependencies
# ==============================
python-dotenv
tiktoken
numpy>=1.26.4
pandas==2.2.3

# ==============================
# LangChain Ecosystem
# (Compatible Versions)
# ==============================
langchain==0.3.3
langchain-core==0.3.14
langchain-community==0.3.2
langchain-text-splitters==0.3.1
langchain-groq==0.2.0

# ==============================
# Graph Database (Neo4j)
# ==============================
neo4j==5.24.0
py2neo==2021.2.4

# ==============================
# Vector Database (FAISS)
# ==============================
faiss-cpu==1.9.0.post1

# ==============================
# Data Retrieval Tools
# ==============================
wikipedia==1.4.0
arxiv==2.1.3

# ==============================
# LLM + API Utilities
# ==============================
openai==1.51.0
requests==2.32.3

# ==============================
# User Interface (Choose One)
# ==============================
# gradio==4.44.1
streamlit==1.39.0

# ==============================
# Notebook / Development Tools
# ==============================
ipykernel
notebook
